refactor: use query instead of repositories
