+++ src/models/query/gateway.py
@@ -1,0 +1,7 @@
+from models.datasources.gatewayapi import GatewayAPI
+
+
+class Gateway:
+    @staticmethod
+    def status():
+        return GatewayAPI.get_status()
