+++ tests/test_query_detail.py
@@ -16,7 +16,7 @@ def test_get(mocker, measure_type):
     from models.query_detail import Detail
 
     m_get: mock.Mock = mocker.patch("lib.query.Query.get")
-    m_insert_detail: mock.Mock = mocker.patch("datasources.database.Database.insert_detail")
+    m_insert_detail: mock.Mock = mocker.patch("models.datasources.database.Database.insert_detail")
     m_get.return_value = MockResponse(
         status_code=200,
         text='{"meter_reading": {"interval_reading": [{"interval_length": "30", '
